removing register fragment and moving the register compose code
