+++ .idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-06-13T21:32:14.666690100Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-06-27T14:43:08.233011900Z" />
   </component>
 </project>
