+++ app/build.gradle
@@ -30,8 +30,8 @@ android {
         applicationId "com.elliottsoftware.calftracker"
         minSdk 23
         targetSdk 33
-        versionCode 13
-        versionName "1.13"
+        versionCode 14
+        versionName "1.14"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
