--- app/src/main/java/com/elliottsoftware/calftracker/presentation/components/register/RegisterFragment.kt
@@ -1,51 +1,0 @@
-package com.elliottsoftware.calftracker.presentation.components.register
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.navigation.findNavController
-import com.elliottsoftware.calftracker.databinding.FragmentRegisterBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-
-/**
- * A simple [Fragment] subclass.
- * Use the [RegisterFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-@AndroidEntryPoint
-class RegisterFragment : Fragment() {
-    private var _binding: FragmentRegisterBinding?=null
-    private val binding get() = _binding!!
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentRegisterBinding.inflate(inflater,container,false)
-        val view = binding.root
-        binding.composeView.apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                RegisterViews(onNavigate = { dest -> findNavController().navigate(dest) })
-            }
-        }
-        return view
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-}
