+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,9 +21,6 @@
             app:popUpTo="@id/loginFragment"
             app:popUpToInclusive="true"
             />
-        <action
-            android:id="@+id/action_loginFragment_to_registerFragment2"
-            app:destination="@id/registerFragment2" />
     </fragment>
     <fragment
         android:id="@+id/mainFragment2"
@@ -53,15 +50,6 @@
             app:popUpTo="@id/mainFragment2"
             app:popUpToInclusive="true" />
 
-    </fragment>
-    <fragment
-        android:id="@+id/registerFragment2"
-        android:name="com.elliottsoftware.calftracker.presentation.components.register.RegisterFragment"
-        android:label="fragment_register"
-        tools:layout="@layout/fragment_register" >
-        <action
-            android:id="@+id/action_registerFragment2_to_mainFragment2"
-            app:destination="@id/mainFragment2" />
     </fragment>
     <fragment
         android:id="@+id/forgotPasswordFragment"
